+++ examples/editor.rs
@@ -98,7 +98,7 @@ pub enum Message {
 }
 
 fn main() {
-    let app = App::default().set_scheme(AppScheme::Gtk);
+    let app = App::default().with_scheme(AppScheme::Gtk);
 
     let (s, r) = channel::<Message>();
     let mut saved = false;